Add helper script to sync jenkins job configs with the release repo
If supplied with a path to a deployment manifest, as well as a path
to a release repo, this script will keep the release repos jenkins
jobs in-sync with whatever is on the remote jenkins master. After
this is run the user can review the changes with ``git diff''
before submitting.

Change-Id: Icf18b0e6474b219fc9f5d51c3d5d7abe6c85a764